NOT FOUND HANDLING
- Add Enter keypress for pop-up notification
- Refactor condition reference of PopUp component
- Send display data from SunInfo, TodayOverview to top Main component
