+++ src/components/SunInfo/index.jsx
@@ -5,11 +5,7 @@ import classes from './SunInfo.module.scss';
 import Skeleton from '../Skeleton';
 
 export default function SunInfo(props) {
-  const { forecastData, loading } = props;
-  const sunInfoList = [
-    { id: 1, name: 'rise', iconSrc: IconSunRise, value: forecastData.sunrise },
-    { id: 2, name: 'set', iconSrc: IconSunSet, value: forecastData.sunset },
-  ];
+  const { loading, sunInfoList } = props;
 
   return (
     <div className="sun-info">
@@ -20,12 +16,12 @@ export default function SunInfo(props) {
             <Skeleton width={300} height={90} bgColor="var(--color-blue-6)" key={info.id} />
           ) : (
             <div className={classes['sun-info__grid']} key={info.id}>
-              <img src={IconSunRise} alt="sunrise" className={classes['sun-info__icon']} />
+              <img src={info.iconSrc} alt="sunrise" className={classes['sun-info__icon']} />
               <p className={classes['sun-info__title']}>
                 Sun
                 <span className={classes['sun-info__title--rise']}>rise</span>
               </p>
-              <p className={classes['sun-info__value']}>{forecastData.sunrise}</p>
+              <p className={classes['sun-info__value']}>{info.value}</p>
             </div>
           ),
         )}
