+++ src/components/WeatherRight/index.jsx
@@ -6,13 +6,13 @@ import SunInfo from '../SunInfo';
 import classes from './WeatherRight.module.scss';
 
 function WeatherRight(props) {
-  const { rainData, forecastData, units, loading } = props;
+  const { forecastData, units, loading, sunInfoList } = props;
   return (
     <div className={classes['weather-right']}>
       <LocationInfo forecastData={forecastData} loading={loading} />
       <BasicInfo forecastData={forecastData} units={units} loading={loading} />
       <RainChanse dailyRainChance={forecastData.dailyRainChance} loading={loading} />
-      <SunInfo forecastData={forecastData} loading={loading} />
+      <SunInfo sunInfoList={sunInfoList} loading={loading} />
     </div>
   );
 }
